Experimental JSON support
Highly experimental support for JSON added.  There are a few
important things to consider with JSON (specifically as it
relates to PHP)

Most importantly, is how we would want data to be decoded.  By
default, PHP will decode JSON values into a stdClass object.
This works if you're checking for that, but will not necissarily
work if you're looking for an array to come back as an array.

For this commit I have decided to decode as associative until I
can determine the best way to handle this.  If we ever decide to
incorporate JSON into the main branch, this will likely need to
become another option set by the user.

In addition, I am attempting to find an example of json_decode_ex,
and json_encode being called (outside of php_json.c iteslf) where
I can test for error conditions.

With this commit, the unit tests are passing with the exception of
object encoding/decoding.