Redis SET 2.8 (round 2)
Redis 2.8 is likely to change to an integer reply for SET commands,
which would currently cause phpredis to return FALSE.

The change is mentioned here:
https://twitter.com/redisfeed/status/298538680220520448

This update uses a new function to handle SET replies, which will
treat either '+OK' or ':1' as a success.

Until the change is confirmed, I will leave this in a feature branch.

Resolves issue #298