Allow user code to catch session exception
This is a somewhat experimental fix to allow php code to catch
an exception when attempting a session_start() call.

The order that PHP appears to call the session hook methods appears
to be as follows:

PS_OPEN_FUNC, PS_READ_FUNC, PS_WRITE_FUNC

Previously, both PS_READ_FUNC and PS_WRITE_FUNC would throw an exception
when Redis was down.  The exception thrown in PS_READ_FUNC can be
caught in a try..catch block but the PS_WRITE_FUNC exception can not.

It's not 100% clear to me why this is, but it appears to be the state
that PHP is in when the PS_WRITE_FUNC is called, causing the exception
to immediately terminate the process.

There is a secondary issue which is that session_start appears to return
true no matter what I return from the read/write hooks.  For that reason
I am still throwing an exception in the read hook, such that a user can
use that to detect a failed session.

Appologies for all the text, but I'm going to use this as a reference.

Addresses issue #325