Initial commit for redis:// scheme support
This commit adds support for using the redis:// scheme protocol to connect to a
redis server, described here: http://www.iana.org/assignments/uri-schemes/prov/redis

There are multiple ways to specify password and database number (both in the uri prefix)
as well as in the query arguments, and the spec is ambiguous here so phpredis will just
use whichever one we parse last (which for us is the query arguments).

In addition, this commit changes $redis->connect() such that it can now be called with
no arguments, in which case phpredis will attempt to connect to a localhost instance
of Redis at the default port of 6379.  I made this change because that's how the redis://
scheme protocol is designed to work, and it should be consistent.