Potential php 7.1 ZINTERSTORE reference fix
Something appears to have changed (more investigation needed) in the way
PHP handles references between 7.0.x and 7.1.x.  This change simply
dereferences values if they are a reference as we iterate through the
weights array.

This fix is very specific to a single array so we probably need to
migrate to using the `zval_get_<type>` heper methods which will
automatically dereference values for us.

Posslbe fix for #1166