Prevent redis cluster from entering inconsistent state
As reported in #1222 phpredis cluster can enter an inconsistent state
when encountering a timeout with persistent connections.

This commit causes redis cluster to close its connection to the node
where the timeout was encountered, to avoid reentering the command
loop out of sync with the server.

It won't prevent failed reads when the timeout happens but the next
time the cluster attempts to read data from that node it will
automatically reconnect.